# new-admin-api/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "ADMIN API",
    "description": "Documentação da API do Novo ADMIN"
  },
  "host": "localhost:3333",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {},
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/aboutUs/{id}": {
      "post": {
        "tags": [
          "AboutUs"
        ],
        "description":"Endpoint para alterar sobre nós",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AboutUs"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          },
          "200": {
            "description": "Ok"
          }
        }
      },
      "delete": {
        "tags":[
          "AboutUs"
        ],
        "description":"Endpoint para excluir sobre nós",
        "parameters":[
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          },
          "200": {
            "description": "Ok"
          }
        }
      }
    },
...
    "/v1/plan": {
      "get": {
        "tags": [
          "Plan"
        ],
        "description": "Endpoints para obter todos os planos",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "$ref": "#/definitions/Plan"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Plan"
        ],
        "description": "Endpoints para criar planos",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Plan"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "$ref": "#/definitions/Plan"
            }
          },
          "400": {
            "description": "Error"
          }
        }
      }
    },
    "/v1/plan/{id}": {
      "post": {
        "tags": [
          "Plan"
        ],
        "description": "Endpoints para alterar planos",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Plan"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "$ref": "#/definitions/Plan"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Plan"
        ],
        "description": "Endpoints para excluir planos",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Ok"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/v1/store": {
      "get": {
        "tags": ["Store"],
        "description":"Endpoint para obter dados da loja",
        "parameters":[],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "$ref": "#/definitions/Store"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": ["Store"],
        "description":"Endpoint para criar loja",
        "parameters":[
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref":"#/definitions/Store"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "$ref": "#/definitions/Store"
            }
          },
          "400": {
            "description": "Error"
          }
        }
      }
    },
    "/v1/store/{id}": {
      "post": {
        "tags": ["Store"],
        "description":"Endpoint para alterar loja",
        "parameters":[
          {
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref":"#/definitions/Store"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "$ref": "#/definitions/Store"
            }
          },
          "400": {
            "description": "Error"
          }
        }
      },
      "delete": {
        "tags": ["Store"],
        "description":"Endpoint para excluir loja",
        "parameters":[
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ok"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "definitions": {
    "AboutUs": {
      "type": "object",
      "properties": {
        "content":{ 
          "type" :"string"
        },
        "user": {
          "type": "string"
        }
      }
    },
...
    "Plan":{
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "price": {
          "type": "number"
        }
      }
    },
...
    "Store":{
      "type": "object",
      "properties": {
        "external_id": {
          "type": "number"
        },
        "name": {
          "type": "string"
        },
        "url": {
          "type": "string"
        },
        "PMC_id": {
          "type": "number"
        },
        "settings": {
          "type": "object"
        }
      }
    }
  }
}